Full refactor of redirect logic
