+++ components/Header/Logo.tsx
@@ -1,8 +1,11 @@
 import Link from "next/link";
 import { useRouter } from "next/router";
 import React from "react";
+import { useEditState } from "tinacms/dist/react";
 
 const Logo = () => {
+  const { edit } = useEditState();
+
   const logo = (
     <span className="main-logo">
       <img
@@ -26,7 +29,11 @@ const Logo = () => {
 
   const onTheMainPage = router.pathname === "/";
 
-  return onTheMainPage ? logo : <Link href="/">{logo}</Link>;
+  return onTheMainPage ? (
+    logo
+  ) : (
+    <Link href={edit ? "/adminHome" : "/"}>{logo}</Link>
+  );
 };
 
 export default Logo;
