--- pages/api/preview/enter.ts
@@ -1,9 +1,0 @@
-import { NextApiRequest, NextApiResponse } from "next";
-
-const handler = async (req: NextApiRequest, res: NextApiResponse) => {
-  res.setDraftMode({ enable: true });
-
-  res.redirect(307, "/admin/index.html");
-};
-
-export default handler;
